pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "latte-pr-agent"
version = "0.1.0"
description = "Enterprise AI Code Review System based on PR-Agent"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "redis>=5.2.0",
    "httpx>=0.27.0",
    "pygithub>=2.5.0",
    "python-gitlab>=5.0.0",
    "openai>=1.55.0",
    "anthropic>=0.42.0",
    "json-repair>=0.30.0",
    "slowapi>=0.1.9",
    "python-multipart>=0.0.17",
    "celery>=5.4.0",
    "tree-sitter>=0.24.0",
    "tree-sitter-python>=0.23.0",
    "pgvector>=0.3.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-httpx>=0.34.0",
    "respx>=0.22.0",
    "testcontainers>=4.8.0",
    "pytest-docker>=3.1.0",
    "aiosqlite>=0.20.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
